задание 7 + правки модулей
