--- js/data-comments.js
@@ -1,41 +1,0 @@
-import {getRandomInteger, createRandomIdFromRangeGenerator, createImageUrl} from './utils.js';
-
-const NAMES = [
-  'Миша',
-  'Артем',
-  'Никита',
-  'Даша',
-  'Акакий',
-  'Алдар Баирович',
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-];
-
-const createRandomComment = (generatorComments, generatorUrl) => ({
-  id: generatorComments(),
-  avatar: createImageUrl(generatorUrl(), 'photos/avatar-', '.svg'),
-  name: NAMES[getRandomInteger(0, NAMES.length - 1)],
-  message: MESSAGES[getRandomInteger(0, MESSAGES.length - 1)],
-});
-
-const createRandomComments = (count) => {
-  const result = [];
-  const commentIdGenerator = createRandomIdFromRangeGenerator(1, count);
-
-  for(let i = 0; i < count; i++) {
-    const urlIdGenerator = createRandomIdFromRangeGenerator(1, 6);
-
-    result.push(createRandomComment(commentIdGenerator, urlIdGenerator));
-  }
-
-  return result;
-};
-
-export {createRandomComments};
